pyproject.toml
[tool.poetry]
name = "wabawachi"
version = "0.1.0"
description = ""
authors = ["chriss060 <alicevravo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-rest-auth = "^0.9.5"
django-rest-authtoken = "^2.1.4"
requests = "^2.28.1"
djangorestframework-api-key = "^2.2.0"
pillow = "^9.3.0"
elasticsearch = "7.17.3"
pyjwt = "^2.6.0"
djongo = "^1.3.6"
numpy = "^1.23.4"
django-cors-headers = "^3.13.0"
mongoengine = "^0.24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
